requirements.txt
Django>=3.0
mysqlclient
djangorestframework==3.11.0